# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is an auto-generated API Docs.",
        "title": "Simeru Scraper API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "jakueenak@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "0.1"
    },
    "basePath": "/",
    "paths": {
        "/api/v1/schedule": {
            "post": {
                "description": "Get schedule based on the request",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Schedule"
                ],
                "summary": "Get Schedule",
                "parameters": [
                    {
                        "description": "Schedule Request",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/rest.ScheduleRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/rest.ScheduleResponseSuccess"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/rest.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/rest.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/study-programs": {
            "get": {
                "description": "Get study programs based on the faculty in the query parameters",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "StudyPrograms"
                ],
                "summary": "Get Study Programs",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Faculty",
                        "name": "faculty",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/rest.StudyProgramsResponseSuccess"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/rest.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/rest.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "entities.RowData": {
            "type": "object",
            "properties": {
                "dosen": {
                    "type": "string"
                },
                "hari": {
                    "type": "string"
                },
                "jam": {
                    "type": "string"
                },
                "kelas": {
                    "type": "string"
                },
                "kode": {
                    "type": "string"
                },
                "matkul": {
                    "type": "string"
                },
                "ruang": {
                    "type": "string"
                },
                "semester": {
                    "type": "string"
                },
                "sks": {
                    "type": "string"
                }
            }
        },
        "entities.StudyPrograms": {
            "type": "object",
            "properties": {
                "faculty": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "value": {
                    "type": "string"
                }
            }
        },
        "rest.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "request_id": {
                    "type": "string"
                }
            }
        },
        "rest.ScheduleRequest": {
            "type": "object",
            "required": [
                "day",
                "study_programs"
            ],
            "properties": {
                "day": {
                    "type": "string",
                    "example": "senin"
                },
                "study_programs": {
                    "type": "string",
                    "maxLength": 255,
                    "minLength": 3,
                    "example": "matematika"
                }
            }
        },
        "rest.ScheduleResponseSuccess": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entities.RowData"
                    }
                }
            }
        },
        "rest.StudyProgramsResponseSuccess": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entities.StudyPrograms"
                    }
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}